refactor(event-stream): extract wall-clock conversion to clock.rs
Pure code motion for #6235. The event_stream/mod.rs god-file fused
wall-clock conversion, producer sequencing, RT_FLOW projection,
connection, replay, drain, and budget ownership in ~2000 lines. Begin
the second-stage split by moving the monotonic-to-wall-clock helpers
into a dedicated clock.rs submodule.

Moved verbatim (no logic or signature change): NS_PER_SEC, the private
read_mono_and_wall_clocks(), and the pub(crate) monotonic_ns_to_unix_ns,
monotonic_ns_to_unix_secs, monotonic_ns_to_unix_secs_subnanos, and
mono_ns_to_wall_clock_unix_ns conversions. read_mono_and_wall_clocks is
widened from private to pub(super) so the RT_FLOW emitters (still in
mod.rs at this step) and the rt_flow tests continue to resolve it.

mod.rs re-exports the four pub(crate) clock fns and imports the private
read helper, so external callers (afxdp/event_emit.rs) and the
event_stream test tree (which reaches them via `use super::*`) resolve
unchanged. cargo build green.
